web: Fix broken github buttons
The github button moved from markdotto.github.com to ghbtns.com, and the old
URL simply 404ed. However, the Github Pages 404 page may not be used within a
frame/iframe, so when JavaScript is enabled, it would display an error message
and remove the outer frame by redirecting to itself. Thus, any visitor to
githubarchive.org with JavaScript enabled would be redirected to a Github Pages
404 page.